mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = ""
authors = [
    {name = "isaac weathers",email = "isaac@isaacweathers.net"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb (>=1.0.9,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "fastapi (==0.115.9)",
    "python-dotenv (>=0.21.0,<1.0.0)" # Added python-dotenv
]

[project.group.dev.dependencies] # Added dev group
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0" # For async HTTP requests in tests

[tool.poetry]
packages = [{include = "mcp_server", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
